pyproject.toml
[tool.poetry]
name = "ine-challenge"
version = "0.1.0"
description = ""
authors = ["Hugo Chavero <hugo.chavero@mercadolibre.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
pytest = "^7.1.3"
requests = "^2.28.1"
pytest-django = "^4.5.2"
pytest-mock = "^3.9.0"
djangorestframework-simplejwt = "^5.2.0"
djangorestframework-api-key = "^2.2.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
pylint = "^2.15.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
